Deprecate 13 utility programs.
mb7k2jstar mb7k2ss mb7kpreprocess mbauvnavusbl mbhsdump
mbhysweeppreprocess mbinsreprocess mbkongsbergpreprocess mbneptune2esf
mbrollbias mbrphsbias mbstripnan mbswplspreprocess

These will still be built.
