+++ .gitignore
@@ -33,13 +33,13 @@ src/mbvelocitytool/mbvelocitytool
 src/otps/mbotps
 src/otps/otps.h
 src/utilities/levitus.h
-src/utilities/mb7k2jstar
-src/utilities/mb7k2ss
-src/utilities/mb7kpreprocess
+src/deprecated/mb7k2jstar
+src/deprecated/mb7k2ss
+src/deprecated/mb7kpreprocess
 src/utilities/mbabsorption
 src/utilities/mbareaclean
 src/utilities/mbauvloglist
-src/utilities/mbauvnavusbl
+src/deprecated/mbauvnavusbl
 src/utilities/mbbackangle
 src/utilities/mbclean
 src/utilities/mbconfig
@@ -55,34 +55,34 @@ src/utilities/mbgetesf
 src/utilities/mbgpstide
 src/utilities/mbgrid
 src/utilities/mbhistogram
-src/utilities/mbhsdump
-src/utilities/mbhysweeppreprocess
+src/deprecated/mbhsdump
+src/deprecated/mbhysweeppreprocess
 src/utilities/mbinfo
-src/utilities/mbinsreprocess
-src/utilities/mbkongsbergpreprocess
+src/deprecated/mbinsreprocess
+src/deprecated/mbkongsbergpreprocess
 src/utilities/mblevitus
 src/utilities/mblist
 src/utilities/mbmakeplatform
 src/utilities/mbminirovnav
 src/utilities/mbmosaic
 src/utilities/mbnavlist
-src/utilities/mbneptune2esf
+src/deprecated/mbneptune2esf
 src/utilities/mbpreprocess
 src/utilities/mbprocess
-src/utilities/mbrollbias
+src/deprecated/mbrollbias
 src/utilities/mbrolltimelag
 src/utilities/mbroutetime
-src/utilities/mbrphsbias
+src/deprecated/mbrphsbias
 src/utilities/mbsegygrid
 src/utilities/mbsegyinfo
 src/utilities/mbsegylist
 src/utilities/mbsegypsd
 src/utilities/mbset
 src/utilities/mbsslayout
-src/utilities/mbstripnan
+src/deprecated/mbstripnan
 src/utilities/mbsvplist
 src/utilities/mbsvpselect
-src/utilities/mbswplspreprocess
+src/deprecated/mbswplspreprocess
 src/utilities/mbtime
 src/utilities/mbtrnpreprocess
 src/utilities/mbvoxelclean
