+++ src/Makefile.in
@@ -155,8 +155,8 @@ am__define_uniq_tagged_files = \
 ETAGS = etags
 CTAGS = ctags
 DIST_SUBDIRS = gsf surf bsio mbtrn mbtrnav mbio mbaux utilities \
-	mbtrnutils gmt otps macros share man html ps mbedit mbnavedit \
-	mbvelocitytool mbview mbgrdviz mbeditviz mbnavadjust
+	deprecated mbtrnutils gmt otps macros share man html ps mbedit \
+	mbnavedit mbvelocitytool mbview mbgrdviz mbeditviz mbnavadjust
 ACLOCAL = @ACLOCAL@
 AMTAR = @AMTAR@
 AM_DEFAULT_VERBOSITY = @AM_DEFAULT_VERBOSITY@
@@ -325,6 +325,7 @@ program_transform_name = @program_transform_name@
 psdir = @psdir@
 pyexecdir = @pyexecdir@
 pythondir = @pythondir@
+runstatedir = @runstatedir@
 sbindir = @sbindir@
 sharedstatedir = @sharedstatedir@
 srcdir = @srcdir@
@@ -340,7 +341,7 @@ top_srcdir = @top_srcdir@
 @BUILD_MBTRN_TRUE@XBUILD_SUB_MBTRNBIN = mbtrnutils
 @BUILD_MBTNAV_TRUE@XBUILD_SUB_MBTNAV = mbtrnav
 SUBDIRS = $(am__append_1) surf bsio $(XBUILD_SUB_MBTRN) \
-	$(XBUILD_SUB_MBTNAV) mbio mbaux utilities \
+	$(XBUILD_SUB_MBTNAV) mbio mbaux utilities deprecated \
 	$(XBUILD_SUB_MBTRNBIN) gmt otps macros share man html ps \
 	$(XBUILD_SUB_MOTIF) $(XBUILD_SUB_GL)
 CLEANFILES =
