+++ src/deprecated/mbstripnan.cc
