+++ src/utilities/Makefile.am
@@ -8,13 +8,9 @@ endif
 
 bin_PROGRAMS =
 bin_PROGRAMS += mbconfig
-bin_PROGRAMS += mb7k2jstar
-bin_PROGRAMS += mb7k2ss
-bin_PROGRAMS += mb7kpreprocess
 bin_PROGRAMS += mbabsorption
 bin_PROGRAMS += mbareaclean
 bin_PROGRAMS += mbauvloglist
-bin_PROGRAMS += mbauvnavusbl
 bin_PROGRAMS += mbbackangle
 bin_PROGRAMS += mbclean
 bin_PROGRAMS += mbcopy
@@ -29,33 +25,24 @@ bin_PROGRAMS += mbgetesf
 bin_PROGRAMS += mbgpstide
 bin_PROGRAMS += mbgrid
 bin_PROGRAMS += mbhistogram
-bin_PROGRAMS += mbhsdump
-bin_PROGRAMS += mbhysweeppreprocess
 bin_PROGRAMS += mbinfo
-bin_PROGRAMS += mbinsreprocess
-bin_PROGRAMS += mbkongsbergpreprocess
 bin_PROGRAMS += mblevitus
 bin_PROGRAMS += mblist
 bin_PROGRAMS += mbmakeplatform
 bin_PROGRAMS += mbminirovnav
 bin_PROGRAMS += mbmosaic
 bin_PROGRAMS += mbnavlist
-bin_PROGRAMS += mbneptune2esf
 bin_PROGRAMS += mbpreprocess
 bin_PROGRAMS += mbprocess
-bin_PROGRAMS += mbrollbias
 bin_PROGRAMS += mbrolltimelag
 bin_PROGRAMS += mbroutetime
-bin_PROGRAMS += mbrphsbias
 bin_PROGRAMS += mbsegygrid
 bin_PROGRAMS += mbsegyinfo
 bin_PROGRAMS += mbsegylist
 bin_PROGRAMS += mbset
 bin_PROGRAMS += mbsslayout
-bin_PROGRAMS += mbstripnan
 bin_PROGRAMS += mbsvplist
 bin_PROGRAMS += mbsvpselect
-bin_PROGRAMS += mbswplspreprocess
 bin_PROGRAMS += mbtime
 bin_PROGRAMS += mbvoxelclean
 bin_PROGRAMS += $(FFTW_PROG)
@@ -86,17 +73,17 @@ LIBS += ${libproj_LIBS}
 LIBS += $(LIBM)
 
 mbconfig_SOURCES = mbconfig.cc
-mb7k2jstar_SOURCES = mb7k2jstar.cc
-mb7k2ss_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mb7k2ss_SOURCES = mb7k2ss.cc
-mb7kpreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mb7kpreprocess_SOURCES = mb7kpreprocess.cc
+# mb7k2jstar_SOURCES = mb7k2jstar.cc
+# mb7k2ss_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mb7k2ss_SOURCES = mb7k2ss.cc
+# mb7kpreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mb7kpreprocess_SOURCES = mb7kpreprocess.cc
 mbabsorption_SOURCES = mbabsorption.cc
 mbareaclean_SOURCES = mbareaclean.cc
 mbauvloglist_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbauvloglist_SOURCES = mbauvloglist.cc
-mbauvnavusbl_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mbauvnavusbl_SOURCES = mbauvnavusbl.cc
+# mbauvnavusbl_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mbauvnavusbl_SOURCES = mbauvnavusbl.cc
 mbbackangle_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbbackangle_SOURCES = mbbackangle.cc
 mbclean_SOURCES = mbclean.cc
@@ -117,13 +104,13 @@ mbgpstide_SOURCES = mbgpstide.cc
 mbgrid_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbgrid_SOURCES = mbgrid.cc
 mbhistogram_SOURCES = mbhistogram.cc
-mbhsdump_SOURCES = mbhsdump.cc
-mbhysweeppreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mbhysweeppreprocess_SOURCES = mbhysweeppreprocess.cc
+# mbhsdump_SOURCES = mbhsdump.cc
+# mbhysweeppreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mbhysweeppreprocess_SOURCES = mbhysweeppreprocess.cc
 mbinfo_SOURCES = mbinfo.cc
-mbinsreprocess_SOURCES = mbinsreprocess.cc
-mbkongsbergpreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mbkongsbergpreprocess_SOURCES = mbkongsbergpreprocess.cc
+# mbinsreprocess_SOURCES = mbinsreprocess.cc
+# mbkongsbergpreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mbkongsbergpreprocess_SOURCES = mbkongsbergpreprocess.cc
 mblevitus_SOURCES = mblevitus.cc
 mblist_SOURCES = mblist.cc
 mbmakeplatform_SOURCES = mbmakeplatform.cc
@@ -131,15 +118,15 @@ mbminirovnav_SOURCES = mbminirovnav.cc
 mbmosaic_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbmosaic_SOURCES = mbmosaic.cc
 mbnavlist_SOURCES = mbnavlist.cc
-mbneptune2esf_SOURCES = mbneptune2esf.cc
+# mbneptune2esf_SOURCES = mbneptune2esf.cc
 mbpreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbpreprocess_SOURCES = mbpreprocess.cc
 mbprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbprocess_SOURCES = mbprocess.cc
-mbrollbias_SOURCES = mbrollbias.cc
+# mbrollbias_SOURCES = mbrollbias.cc
 mbrolltimelag_SOURCES = mbrolltimelag.cc
 mbroutetime_SOURCES = mbroutetime.cc
-mbrphsbias_SOURCES = mbrphsbias.cc
+# mbrphsbias_SOURCES = mbrphsbias.cc
 mbsegygrid_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbsegygrid_SOURCES = mbsegygrid.cc
 mbsegyinfo_SOURCES = mbsegyinfo.cc
@@ -147,11 +134,11 @@ mbsegylist_SOURCES = mbsegylist.cc
 mbset_SOURCES = mbset.cc
 mbsslayout_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
 mbsslayout_SOURCES = mbsslayout.cc
-mbstripnan_SOURCES = mbstripnan.cc
+# mbstripnan_SOURCES = mbstripnan.cc
 mbsvplist_SOURCES = mbsvplist.cc
 mbsvpselect_SOURCES = mbsvpselect.cc geodesic.c
-mbswplspreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
-mbswplspreprocess_SOURCES = mbswplspreprocess.cc
+# mbswplspreprocess_LDADD = ${top_builddir}/src/mbaux/libmbaux.la
+# mbswplspreprocess_SOURCES = mbswplspreprocess.cc
 mbtime_SOURCES = mbtime.cc
 mbvoxelclean_SOURCES = mbvoxelclean.cc
 if BUILD_FFTW
