+++ test/deprecated/Makefile.am
@@ -1,0 +1,49 @@
+TEST_EXTENSIONS = .py
+
+TESTS =
+
+# if HAVE_PYTHON
+check_SCRIPTS =
+
+TESTS += mb7k2jstar_test.py
+check_SCRIPTS += mb7k2jstar_test.py
+
+TESTS += mb7k2ss_test.py
+check_SCRIPTS += mb7k2ss_test.py
+
+TESTS += mb7kpreprocess_test.py
+check_SCRIPTS += mb7kpreprocess_test.py
+
+TESTS += mbauvnavusbl_test.py
+check_SCRIPTS += mbauvnavusbl_test.py
+
+TESTS += mbhsdump_test.py
+check_SCRIPTS += mbhsdump_test.py
+
+TESTS += mbhysweeppreprocess_test.py
+check_SCRIPTS += mbhysweeppreprocess_test.py
+
+TESTS += mbinsreprocess_test.py
+check_SCRIPTS += mbinsreprocess_test.py
+
+TESTS += mbkongsbergpreprocess_test.py
+check_SCRIPTS += mbkongsbergpreprocess_test.py
+
+TESTS += mbneptune2esf_test.py
+check_SCRIPTS += mbneptune2esf_test.py
+
+# TODO(schwehr): Why is this not built?
+# TESTS += mbps_test.py
+# check_SCRIPTS += mbps_test.py
+
+TESTS += mbrollbias_test.py
+check_SCRIPTS += mbrollbias_test.py
+
+# TESTS += mbstripnan_test.py
+# check_SCRIPTS += mbstripnan_test.py
+
+TESTS += mbswplspreprocess_test.py
+check_SCRIPTS += mbswplspreprocess_test.py
+
+## HAVE_PYTHON
+# endif
