+++ test/utilities/Makefile.am
@@ -5,15 +5,6 @@ TESTS =
 # if HAVE_PYTHON
 check_SCRIPTS =
 
-TESTS += mb7k2jstar_test.py
-check_SCRIPTS += mb7k2jstar_test.py
-
-TESTS += mb7k2ss_test.py
-check_SCRIPTS += mb7k2ss_test.py
-
-TESTS += mb7kpreprocess_test.py
-check_SCRIPTS += mb7kpreprocess_test.py
-
 TESTS += mbabsorption_test.py
 check_SCRIPTS += mbabsorption_test.py
 
@@ -22,9 +13,6 @@ check_SCRIPTS += mbareaclean_test.py
 
 TESTS += mbauvloglist_test.py
 check_SCRIPTS += mbauvloglist_test.py
-
-TESTS += mbauvnavusbl_test.py
-check_SCRIPTS += mbauvnavusbl_test.py
 
 TESTS += mbbackangle_test.py
 check_SCRIPTS += mbbackangle_test.py
@@ -71,20 +59,8 @@ check_SCRIPTS += mbgrid_test.py
 TESTS += mbhistogram_test.py
 check_SCRIPTS += mbhistogram_test.py
 
-TESTS += mbhsdump_test.py
-check_SCRIPTS += mbhsdump_test.py
-
-TESTS += mbhysweeppreprocess_test.py
-check_SCRIPTS += mbhysweeppreprocess_test.py
-
 TESTS += mbinfo_test.py
 check_SCRIPTS += mbinfo_test.py
-
-TESTS += mbinsreprocess_test.py
-check_SCRIPTS += mbinsreprocess_test.py
-
-TESTS += mbkongsbergpreprocess_test.py
-check_SCRIPTS += mbkongsbergpreprocess_test.py
 
 TESTS += mblevitus_test.py
 check_SCRIPTS += mblevitus_test.py
@@ -104,30 +80,17 @@ check_SCRIPTS += mbmosaic_test.py
 TESTS += mbnavlist_test.py
 check_SCRIPTS += mbnavlist_test.py
 
-TESTS += mbneptune2esf_test.py
-check_SCRIPTS += mbneptune2esf_test.py
-
 TESTS += mbpreprocess_test.py
 check_SCRIPTS += mbpreprocess_test.py
 
 TESTS += mbprocess_test.py
 check_SCRIPTS += mbprocess_test.py
 
-# TODO(schwehr): Why is this not built?
-# TESTS += mbps_test.py
-# check_SCRIPTS += mbps_test.py
-
-TESTS += mbrollbias_test.py
-check_SCRIPTS += mbrollbias_test.py
-
 TESTS += mbrolltimelag_test.py
 check_SCRIPTS += mbrolltimelag_test.py
 
 TESTS += mbroutetime_test.py
 check_SCRIPTS += mbroutetime_test.py
-
-TESTS += mbrphsbias_test.py
-check_SCRIPTS += mbrphsbias_test.py
 
 TESTS += mbsegygrid_test.py
 check_SCRIPTS += mbsegygrid_test.py
@@ -147,17 +110,11 @@ check_SCRIPTS += mbset_test.py
 TESTS += mbsslayout_test.py
 check_SCRIPTS += mbsslayout_test.py
 
-# TESTS += mbstripnan_test.py
-# check_SCRIPTS += mbstripnan_test.py
-
 TESTS += mbsvplist_test.py
 check_SCRIPTS += mbsvplist_test.py
 
 TESTS += mbsvpselect_test.py
 check_SCRIPTS += mbsvpselect_test.py
-
-TESTS += mbswplspreprocess_test.py
-check_SCRIPTS += mbswplspreprocess_test.py
 
 TESTS += mbtime_test.py
 check_SCRIPTS += mbtime_test.py
